refactor(db): 修改hotel_rating表的外键为integer类型

+++ backend/base/src/models/hotel_rating.rs
@@ -7,8 +7,8 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
-    pub user_id: i64,
-    pub hotel_id: i64,
+    pub user_id: i32,
+    pub hotel_id: i32,
     pub time: DateTimeWithTimeZone,
     #[sea_orm(column_type = "Decimal(Some((2, 1)))")]
     pub rating: Decimal,
